requirements.txt
python-slugify
myst-parser
